Move MyClients to /me/clients and ad a general /clients page
